chore: 🤖 create related sessions on connect in dev for mocks
